refactor: extract beauty-specific code from central files
Move beauty/consulting logic out of index.ts, ChatDrawer.ts, and
renderUISpec.ts into dedicated modules to minimize PR diff in central
files and preserve backward compatibility for non-beauty clients.

New files:
- assistant-mode.ts: AssistantMode type, parseRedirectMode, asRecord/firstString
- ConsultingStylePicker.ts: style variation picker (~160 lines from renderProductGrid)

Enriched component files:
- PhotoAnalysisCard.ts: parsePhotoAnalysisProps, renderPhotoAnalysisBubble
- BeautyPhotoStep.ts: parseBeautyPhotoStepProps, BeautyPhotoStepProps

Backward-compat fixes:
- meta.assistantMode only sent when mode != shopping (no wire change for existing clients)
- thinkingMessages truncation gated by non-shopping mode only
- debugLog tags fixed: 'beauty' -> 'mode'/'chat'/'redirect'

Net: -162 lines from central files.
